+++ js/exports.js
@@ -1,0 +1,3 @@
+var keepnote = require('./keepnote.jsx');
+
+window.KeepNoteApp = keepnote.KeepNoteApp;
