+++ js/treeview_raw.jsx
@@ -1,0 +1,162 @@
+var NotebookTreeRaw = React.createClass({
+    getInitialState: function () {
+        var node = this.props.node;
+        var expanded = node.get("expanded") || false;
+        return {
+            expanded: expanded,
+            filesExpanded: false
+        };
+    },
+
+    render: function() {
+        var node = this.props.node;
+
+        // Get children.
+        var children = [];
+        for (var i=0; i<node.children.length; i++) {
+            var child = node.children[i];
+            if (child.fetched) {
+                children.push(
+                  <li key={child.id}>
+                   <NotebookTreeRaw
+                    node={child}
+                    currentNode={this.props.currentNode}
+                    onViewNode={this.props.onViewNode} /></li>);
+            }
+        }
+
+        var displayChildren = (this.state.expanded ? "inline" : "none");
+        var displayFiles = (this.state.filesExpanded ? "block" : "none");
+        var nodeClass = "node-tree";
+        if (node == this.props.currentNode)
+            nodeClass += " active";
+
+        // Build node title.
+        var title = <span className="title">{node.get('title')}</span>;
+        if (node.isPage()) {
+            // Notebook page.
+            title = <a href="#" onClick={this.onPageClick}>{title}</a>;
+        } else if (node.get("payload_filename")) {
+            // Attached file.
+            title = <a href={node.payloadUrl()} target="_blank">{title}</a>;
+        }
+
+        return <div className={nodeClass}>
+          <a className="expand" onClick={this.toggleChildren} href="#">+</a>
+          {title}
+          [<a href={node.url()}>attr</a>]&nbsp;
+          [<a onClick={this.toggleFiles} href="#">files</a>]
+
+          <div className="files" style={{display: displayFiles}}>
+            <NotebookFile
+             file={node.file}
+             showFilename={false}
+             expanded={true} />
+          </div>
+
+          <div className="children" style={{display: displayChildren}}>
+            <ul>
+              {children}
+            </ul>
+          </div>
+        </div>;
+    },
+
+    toggleChildren: function (e) {
+        e.preventDefault();
+
+        var expanded = !this.state.expanded;
+        this.setState({expanded: expanded});
+        if (expanded) {
+            this.props.node.fetchChildren();
+        }
+    },
+
+    toggleFiles: function (e) {
+        e.preventDefault();
+
+        var expanded = !this.state.filesExpanded;
+        this.setState({filesExpanded: expanded});
+        if (expanded)
+            this.props.node.file.fetch();
+    },
+
+    onPageClick: function (e) {
+        e.preventDefault();
+
+        if (this.props.onViewNode)
+            this.props.onViewNode(this.props.node);
+    }
+});
+
+
+var NotebookFile = React.createClass({
+    getDefaultProps: function () {
+        return {
+            showFilename: true,
+            expanded: false
+        };
+    },
+
+    getInitialState: function () {
+        return {
+            expanded: this.props.expanded
+        };
+    },
+
+    render: function () {
+        var file = this.props.file;
+
+        // Populate child list.
+        var children = [];
+        var fileChildren = file.children;
+        for (var i=0; i<fileChildren.length; i++) {
+            var child = fileChildren[i];
+            children.push(<li key={i}><NotebookFile file={child} /></li>);
+        }
+
+        var displayChildren = (this.state.expanded ? "block" : "none");
+
+        var filenameNode = null;
+        if (this.props.showFilename) {
+            // Render filename link.
+            var filename = file.basename() + (file.isDir ? '/' : '');
+
+            var onClick = null;
+            var href = "#";
+            if (file.isDir) {
+                onClick = this.toggleChildren;
+            } else {
+                href = file.url();
+            }
+
+            filenameNode = <a key="0" className="filename"
+                            onClick={onClick} href={href}>{filename}</a>;
+        }
+
+        return <div style={this.props.style}>
+            {filenameNode}
+            <ul style={{display: displayChildren}}>
+              {children}
+            </ul>
+          </div>;
+    },
+
+    toggleChildren: function (e) {
+        e.preventDefault();
+
+        var expanded = !this.state.expanded;
+        this.setState({expanded: expanded});
+        if (expanded)
+            this.props.file.fetch();
+    }
+});
+
+
+// Exports.
+if (typeof(module) !== 'undefined') {
+    module.exports = {
+        NotebookTreeRaw: NotebookTreeRaw
+        NotebookFile: NotebookFile
+    };
+}
